be35555f-afad-4385-a24f-21bcc0000f25/requirements.txt
discord.py
imgurpython
flask
dotenv